+++ src/controllers/event.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interfaces";
 import response from "../utils/response";
 import EventModel, { eventDAO, TEvent } from "../models/event.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -54,7 +54,20 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(
+          res,
+          "Failed to get one event. Id is not valid"
+        );
+      }
+
       const result = await EventModel.findById(id);
+      if (!result) {
+        return response.notFound(
+          res,
+          "Failed to get one event. Event is not found"
+        );
+      }
       response.success(res, result, "Success to find one event");
     } catch (error) {
       response.error(res, error, "Failed to find one event");
@@ -63,6 +76,13 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(
+          res,
+          "Failed to update event. Id is not valid"
+        );
+      }
+
       const result = await EventModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -74,6 +94,13 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(
+          res,
+          "Failed to remove event. Id is not valid"
+        );
+      }
+
       const result = await EventModel.findByIdAndDelete(id, { new: true });
       response.success(res, result, "Success to remove an event");
     } catch (error) {
@@ -84,6 +111,12 @@ export default {
     try {
       const { slug } = req.params;
       const result = await EventModel.findOne({ slug });
+      if (!result) {
+        return response.notFound(
+          res,
+          "Failed to find an event by slug. Event is not found"
+        );
+      }
       response.success(res, result, "Success to find an event by slug");
     } catch (error) {
       response.error(res, error, "Failed to find an event by slug");
